requirements.txt
agno==1.5.0
anthropic==0.51.0
certifi==2024.6.2
fast_flights==2.2
fastapi==0.110.0
gradio==5.42.0
pandas==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
uvicorn==0.35.0
